endereços de variáveis

+++ endereco-de-variaveis/endereco02.c
@@ -5,8 +5,8 @@ int main() {
     char a = '0';
     short b = 1;
     int c = 3;
-    long d = 4 ; 
-    long e = 5;
+    long d = 4; 
+    double e = 5.;
 
      // & obtem o endereço de memória das váriaveis.
      // hexadecimal | decimal | conteúdo
@@ -14,7 +14,7 @@ int main() {
     printf("&b = %p | &b = %ld | b = %d\n", &b, &b, b);
     printf("&c = %p | &c = %ld | c = %d\n", &c, &c, c);
     printf("&d = %p | &d = %ld | d = %d\n", &d, &d, d);
-    printf("&e = %p | &d = %ld | e = %d\n\n", &e, &e, e);
+    printf("&e = %p | &d = %ld | e = %lf\n\n", &e, &e, e);
   
 
     return 0;
